+++ pyfda/input_widgets/input_coeffs_ui.py
@@ -41,7 +41,7 @@ class Input_Coeffs_UI(QWidget):
         self.cmb_qfrmt_items = [
             "<span>Quantization format for coefficients (affects only "
             "the display, not the stored values).</span>",
-            ('float', "Float64", "<span>Full precision floating point format</span>"),
+            ('float64', "Float64", "<span>Full precision floating point format</span>"),
             ('float32', "Float32", "<span>Single precision floating point format</span>"),
             ('qint', "Integer", "<span>Integer format with <i>WI</i> + 1 bits "
              "(range -2<sup>WI</sup> ... 2<sup>WI</sup> - 1)</span>"),
@@ -49,7 +49,7 @@ class Input_Coeffs_UI(QWidget):
              "<span>General fractional format with <i>WI</i> + <i>WF</i> + 1 bits "
              "(range -2<sup>WI</sup> ... 2<sup>WI</sup> - 2<sup>WF</sup>).</span>")
             ]
-        self.cmb_qfrmt_default = 'float'
+        self.cmb_qfrmt_default = 'float64'
 
         self.cmb_fx_base_items = [
             "<span>Select the coefficient fixpoint display format.</span>",
